refactor: response dto로 이동
Description
service와 domain간 dto 적용
