+++ src/main/java/com/dimsssss/raid/raid/application/RaidRecordService.java
@@ -1,6 +1,9 @@
 package com.dimsssss.raid.raid.application;
 
 import com.dimsssss.raid.raid.domain.*;
+import com.dimsssss.raid.raid.domain.dto.BossStateResponseDto;
+import com.dimsssss.raid.raid.domain.dto.RaidStartResponseDto;
+import com.dimsssss.raid.raid.domain.dto.RankingResponseDto;
 import com.dimsssss.raid.raid.presentation.dto.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
