+++ src/main/java/com/dimsssss/raid/raid/presentation/RaidRecordController.java
@@ -2,6 +2,9 @@ package com.dimsssss.raid.raid.presentation;
 
 import com.dimsssss.raid.raid.application.RaidRecordService;
 import com.dimsssss.raid.raid.domain.RaidTimeoutException;
+import com.dimsssss.raid.raid.domain.dto.BossStateResponseDto;
+import com.dimsssss.raid.raid.domain.dto.RaidStartResponseDto;
+import com.dimsssss.raid.raid.domain.dto.RankingResponseDto;
 import com.dimsssss.raid.raid.presentation.dto.*;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
